The CountAllWalls() function disagrees with the print walls function, which both disagree with the actual map being generated. Something is very wrong.

+++ graphics.h
@@ -55,8 +55,8 @@ typedef struct _Node{
 } Node;
 
 typedef struct _GenerationInfo{
-    int spawnChanceModifier;
-    int spawnChance;
+    float spawnChanceModifier;
+    float spawnChance;
 
     int creationAttempts;
     int wallsCreated;
